requirements.txt
beautifulsoup4==4.12.3
discord.py>=2.4.0
audioop-lts==0.2.1
python-dotenv==1.0.0
selenium==4.15.2
pytest==7.4.3
pytest-mock==3.12.0
mypy==1.5.1
flake8==6.0.0